Prefere Play's FakeKeyStore and remove duplicate code
